==> code/03_prediction/01_summary.ipynb <==
# Create summary data for signatures

Measure median and IQR for each variable and store them as Parquets.

Data available at https://github.com/urbangrammarai/signatures_gb.

import dask.dataframe as dd
import pandas as pd

data_folder = "/Users/martin/Library/CloudStorage/OneDrive-SharedLibraries-TheAlanTuringInstitute/Daniel Arribas-Bel - demoland_data"

Define a key to match data.

form_key = {
    "area of building": "sdbAre",
    "courtyard area of building": "sdbCoA",
    "circular compactness of building": "ssbCCo",
    "corners of building": "ssbCor",
    "squareness of building": "ssbSqu",
    "equivalent rectangular index of building": "ssbERI",
    "centroid - corner mean distance of building": "ssbCCM",
    "centroid - corner distance deviation of building": "ssbCCD",
    "orientation of building": "stbOri",
    "area of ETC": "sdcAre",
    "circular compactness of ETC": "sscCCo",
    "equivalent rectangular index of ETC": "sscERI",
    "covered area ratio of ETC": "sicCAR",
    "cell alignment of building": "stbCeA",
    "alignment of neighbouring buildings": "mtbAli",
    "mean distance between neighbouring buildings": "mtbNDi",
    "perimeter-weighted neighbours of ETC": "mtcWNe",
    "mean inter-building distance": "ltbIBD",
    "width of street profile": "sdsSPW",
    "width deviation of street profile": "sdsSWD",
    "openness of street profile": "sdsSPO",
    "length of street segment": "sdsLen",
    "linearity of street segment": "sssLin",
    "mean segment length within 3 steps": "ldsMSL",
    "node degree of junction": "mtdDeg",
    "local proportion of 3-way intersections of street network": "linP3W",
    "local proportion of 4-way intersections of street network": "linP4W",
    "local proportion of cul-de-sacs of street network": "linPDE",
    "local closeness of street network": "lcnClo",
    "local cul-de-sac length of street network": "ldsCDL",
    "square clustering of street network": "xcnSCl",
    "local degree weighted node density of street network": "linWID",
    "street alignment of building": "stbSAl",
    "area covered by edge-attached ETCs": "sdsAre",
    "buildings per meter of street segment": "sisBpM",
    "reached ETCs by neighbouring segments": "misCel",
    "reached ETCs by tessellation contiguity": "ltcRea",
    "area of enclosure": "ldeAre",
    "circular compactness of enclosure": "lseCCo",
    "equivalent rectangular index of enclosure": "lseERI",
    "orientation of enclosure": "lteOri",
    "perimeter-weighted neighbours of enclosure": "lteWNB",
    "area-weighted ETCs of enclosure": "lieWCe",
}

Read form.

form = (
    dd.read_parquet("signatures_gb/form", columns=["hindex"] + list(form_key))
    .compute()
    .set_index("hindex")
    .rename(columns=form_key)
)

Read signature labels

signature_type = (
    dd.read_parquet("urbangrammar_samba/spatial_signatures/signatures/hindex_to_type")
    .compute()
    .set_index("hindex")
)

Sort and merge via index.

form = form.sort_index()
signature_type = signature_type.sort_index()
form = form.merge(
    signature_type[["type"]], how="left", left_index=True, right_index=True
)

form.head()

Group by a signature type, get quartiles and IQR.

grouper = form.groupby("type")

median = grouper.quantile(0.5)
q25 = grouper.quantile(0.25)
q75 = grouper.quantile(0.75)
iqr = q75 - q25

Properly name types.

renamer = {
    "Warehouse land": "Warehouse/Park land",
    "Hyper distilled urbanity": "Hyper concentrated urbanity",
    "Distilled urbanity": "Concentrated urbanity",
}
median = median.rename(index=renamer)
iqr = iqr.rename(index=renamer)

median.to_parquet(f"{data_folder}/sampling/median_form.parquet")
iqr.to_parquet(f"{data_folder}/sampling/iqr_form.parquet")

## Non-morphological data

Key:

fn_key = {
    "Workplace population [Agriculture, energy and water]": "A, B, D, E. Agriculture, energy and water",
    "Workplace population [Manufacturing]": "C. Manufacturing",
    "Workplace population [Construction]": "F. Construction",
    "Workplace population [Distribution, hotels and restaurants]": "G, I. Distribution, hotels and restaurants",
    "Workplace population [Transport and communication]": "H, J. Transport and communication",
    "Workplace population [Financial, real estate, professional and administrative activities]": "K, L, M, N. Financial, real estate, professional and administrative activities",
    "Workplace population [Public administration, education and health]": "O,P,Q. Public administration, education and health",
    "Workplace population [Other]": "R, S, T, U. Other",
}

Read and merge labels (and areas).

function = (
    dd.read_parquet(
        "signatures_gb/function",
        columns=[
            "hindex",
            "Population",
            "Land cover [Non-irrigated arable land]",
            "Land cover [Industrial or commercial units]",
            "Land cover [Sport and leisure facilities]",
            "Land cover [Green urban areas]",
            "Land cover [Discontinuous urban fabric]",
            "Land cover [Pastures]",
            "Land cover [Continuous urban fabric]",
        ]
        + list(fn_key),
    )
    .compute()
    .set_index("hindex")
)

function = function.merge(
    signature_type[["type"]], how="left", left_index=True, right_index=True
).merge(form[["sdcAre"]], how="left", left_index=True, right_index=True)
function = function.rename(columns=fn_key)

Normalize subset by area.

subset = [
    "Population",
    "A, B, D, E. Agriculture, energy and water",
    "C. Manufacturing",
    "F. Construction",
    "G, I. Distribution, hotels and restaurants",
    "H, J. Transport and communication",
    "K, L, M, N. Financial, real estate, professional and administrative activities",
    "O,P,Q. Public administration, education and health",
    "R, S, T, U. Other",
]
function[subset] = function[subset].divide(function.sdcAre, axis=0)

Group by type, get quartiles and IQR.

grouper = function.groupby("type")
median = grouper.quantile(0.5)
q25 = grouper.quantile(0.25)
q75 = grouper.quantile(0.75)
iqr = q75 - q25
median = median.rename(index=renamer)
iqr = iqr.rename(index=renamer)

Save.

median.drop(columns="sdcAre").to_parquet(
    f"{data_folder}/sampling/median_function.parquet"
)
iqr.drop(columns="sdcAre").to_parquet(f"{data_folder}/sampling/iqr_function.parquet")
